+++ CHANGELOG.md
@@ -3,3 +3,9 @@
     * stabilized initial design
     * make more palatable to linux
     * change licenseing to GPL
+
+* 0.2 - 9/28/2019 - bug fixes
+
+    * honor dark ui preference
+    * don't fail if notification service is not available
+    * use pcmanfm when session is LXDE-pi